Section 14: House position calculations

Calculate house position of a given body.

swe_house_pos(armc, geolat, eps, hsys, xpin)

swe_gauquelin_sector(
  jd_ut,
  ipl,
  starname,
  ephe_flag,
  imeth,
  geopos,
  atpress,
  attemp
)

Arguments

armc

right ascension of the MC as double (deg)

geolat

geographic latitude as double (deg)

eps

ecliptic obliquity as double (deg)

hsys

house method, one-letter case sensitive as char

xpin

longitude and latitude of the given body as numeric vector (deg)

jd_ut

UT Julian day number as double (day)

ipl

Body/planet as integer (SE$SUN=0, SE$MOON=1, ... SE$PLUTO=9)

starname

Star name as string ("" for no star)

ephe_flag

Ephemeris flag as integer (SE$FLG_JPLEPH=1, SE$FLG_SWIEPH=2 or SE$FLG_MOSEPH=4)

imeth

Gauquelin method as integer (0, 1, 2, 3, 4 or 5)

geopos

position as numeric vector (longitude, latitude, height)

atpress

Atmospheric pressure as double (hPa)

attemp

Atmospheric temperature as double (Celsius)

Value

swe_house_pos returns a list with named entries: return how far from body's cusp as double, and serr error message as string.

swe_gauquelin_sector returns a list with named entries: return status flag as integer, dgsect for Gauquelin sector as double and serr error message as string

Details

swe_house_pos()

Calculate house position of given body.

swe_gauquelin_sector()

Compute the Gauquelin sector position of a planet or star.

See also

Examples

swe_house_pos(12, 53, 23, 'B', c(0,0))
#> $return
#> [1] 9.435262
#> 
#> $serr
#> [1] ""
#> 
data(SE)
swe_gauquelin_sector(1234567.5,SE$VENUS,"",SE$FLG_MOSEPH,0,c(0,50,10),1013.25,15)
#> $return
#> [1] 0
#> 
#> $dgsect
#> [1] 31.41367
#> 
#> $serr
#> [1] ""
#>