Section 4: Fixed stars functions

The following functions are used to calculate positions of fixed stars.

swe_fixstar2_mag(starname)

swe_fixstar2(starname, jd_et, iflag)

swe_fixstar2_ut(starname, jd_ut, iflag)

Arguments

starname
Star name as string ("" for no star)

jd_et
ET Julian day number as double (day)

iflag
Calculation flag as integer, many options possible (section 2.3)

jd_ut
UT Julian day number (day)

Value

swe_fixstar2_mag returns a list with named entries: return status flag as integer, starname updated star name as string, mag magnitude of star as double, and serr for error message as string.

swe_fixstar2 returns a list with named entries: return status flag as integer, starname updated star name as string, xx star phenomena as numeric vector, and serr error message as string.

swe_fixstar2_ut returns a list with named entries: return status flag as integer, starname updated star name as string, xx star information as numeric vector, and serr for error message as string.

Details

swe_fixstar2_mag()

Calculate visible magnitude (Vmag) of star.

swe_fixstar2()

Compute information of star using ET.

swe_fixstar2_ut()

Compute information of star using UT

See also

Examples

data(SE)
swe_fixstar2_mag("sirius")

#> $return
#> [1] 0
#> 
#> $starname
#> [1] "Sirius,alCMa"
#> 
#> $mag
#> [1] -1.46
#> 
#> $serr
#> [1] ""
#> 

swe_set_topo(0,50,10)
swe_fixstar2("sirius",1234567,SE$FLG_TOPOCTR+SE$FLG_MOSEPH+SE$FLG_EQUATORIAL)

#> $return
#> [1] 34820
#> 
#> $starname
#> [1] "Sirius,alCMa"
#> 
#> $xx
#> [1]     64.65734    -17.80520 547924.62574      0.00000      0.00000
#> [6]      0.00000
#> 
#> $serr
#> [1] ""
#> 

swe_fixstar2_ut("sirius",1234567,SE$FLG_TOPOCTR+SE$FLG_MOSEPH+SE$FLG_EQUATORIAL)

#> $return
#> [1] 34820
#> 
#> $starname
#> [1] "Sirius,alCMa"
#> 
#> $xx
#> [1]     64.65729    -17.80518 547924.62843      0.00000      0.00000
#> [6]      0.00000
#> 
#> $serr
#> [1] ""
#>