Section 1: The Ephemeris file related functions

Several initialization functions

swe_set_ephe_path(path)

swe_close()

swe_set_jpl_file(fname)

swe_version()

swe_get_library_path()

Arguments

path
Directory for the sefstars.txt, swe_deltat.txt and jpl files

fname
JPL ephemeris name as string (JPL ephemeris file, e.g. de431.eph)

Value

swe_version returns Swiss Ephemeris software version as string

swe_get_library_path returns the path in which the executable resides as string

Details

swe_set_ephe_path()

This is the first function that should be called before any other function of the Swiss Ephemeris. Even if you don't want to set an ephemeris path and use the Moshier ephemeris, it is nevertheless recommended to call swe_set_ephe_path(NULL), because this function makes important initializations. If you don't do that, the Swiss Ephemeris may work, but the results may be not 100% consistent.

swe_close()

At the end of your computations this function releases most resources (open files and allocated memory) used by Swiss Ephemeris.

swe_set_jpl_file()

Set name of JPL ephemeris file.

swe_version()

The function provides the version number of the Swiss Ephemeris software.

swe_get_library_path()

The function provides the path where the executable resides.

See also

Examples

if (FALSE) swe_set_ephe_path("c:\\sweph\\ephe")
swe_close()
swe_set_jpl_file("de431.eph")
swe_version()

#> [1] "2.08"

swe_get_library_path()

#> [1] "/home/travis/R/Library/swephR/libs/swephR.so"