Section 10: Sidereal mode functions

Functions to support the determination of sidereal information

swe_set_sid_mode(sid_mode, t0, ayan_t0)

swe_get_ayanamsa_name(sid_mode)

swe_get_ayanamsa_ex_ut(jd_ut, iflag)

swe_get_ayanamsa_ex(jd_et, iflag)

Arguments

sid_mode
Sidereal mode as integer

t0
Reference date as double (day)

ayan_t0
The initial latitude value of the ayanamsa as double (deg)

jd_ut
UT Julian day number as double (day)

iflag
Computation flag as integer, many options possible (section 2.3,1)

jd_et
ET Julian day number as double (day)

Value

swe_get_ayanamsa_name returns name of ayanamsa method as string

swe_get_ayanamsa_ex_ut returns a list with named entries: return status flag as integer, daya ayanamsa value as double and serr error message as string

swe_get_ayanamsa_ex returns a list with named entries: return status flag as integer, daya ayanamsa value as double and serr error message as string

Details

swe_set_sid_mode()

Set the mode for sidereal computations.

swe_get_ayanamsa_name()

Get the mode name for sidereal computations.

swe_get_ayanamsa_ex_ut()

It computes ayanamsa using UT.

swe_get_ayanamsa_ex()

It computes ayanamsa using ET.

See also

Examples

data(SE)
swe_set_sid_mode(SE$SIDM_FAGAN_BRADLEY,0,0)
swe_get_ayanamsa_name(SE$SIDM_FAGAN_BRADLEY)

#> [1] "Fagan/Bradley"

swe_get_ayanamsa_ex_ut(2458346.82639,SE$FLG_MOSEPH)

#> $return
#> [1] 68
#> 
#> $daya
#> [1] 24.99688
#> 
#> $serr
#> [1] ""
#> 

swe_get_ayanamsa_ex(2458346.82639,SE$FLG_MOSEPH)

#> $return
#> [1] 68
#> 
#> $daya
#> [1] 24.99688
#> 
#> $serr
#> [1] ""
#>