Section 13: House cusp, ascendant and Medium Coeli calculations

Calculate house cusp, ascendant, Medium Coeli, etc. calculations

swe_houses_ex(jd_ut, cuspflag, geolat, geolon, hsys)

swe_houses_armc(armc, geolat, eps, hsys)

swe_house_name(hsys)

Arguments

jd_ut

UT Julian day number as double (day)

cuspflag

cusp flag as integer (0 [tropical], SE$FLG_SIDEREAL, SE$FLG_RADIANS)

geolat

geographic latitude as double (deg)

geolon

geographic longitude as double (deg)

hsys

house method, one-letter case sensitive as char

armc

right ascension of the MC as double (deg)

eps

ecliptic obliquity as double (deg)

Value

swe_houses_ex returns a list with named entries: return status flag as integer,

cusps cusps values as double and ascmc ascendent, MCs. etc. as double.

swe_houses_armc returns a list with named entries: return status flag as integer,

cusps cusps values as double and ascmc ascendent, MCs, etc. as double.

swe_house_name returns the house name as string

Details

swe_houses_ex()

Calculate houses' cusps, ascendant, Medium Coeli (MC), etc.

swe_houses_armc()

Calculate houses' information from the right ascension of the Medium Coeli (MC).

swe_houses_name()

Provide the house name.

See also

Examples

swe_houses_ex(1234567, 0, 53, 0, 'B')
#> $return
#> [1] 0
#> 
#> $cusps
#>  [1]   0.00000  46.64518  67.89822  88.19583 108.40898 145.68710 186.61910
#>  [8] 226.64518 247.89822 268.19583 288.40898 325.68710   6.61910   0.00000
#> [15]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [22]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [29]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [36]   0.00000   0.00000
#> 
#> $ascmc
#>  [1]  46.64518 288.40898 289.99675 196.36986  21.69681  13.76371  31.65048
#>  [8] 193.76371   0.00000   0.00000
#> 
swe_houses_armc(12, 53, 23, 'B')
#> $return
#> [1] 0
#> 
#> $cusps
#>  [1]   0.00000 125.97068 147.38669 169.96807 193.00244 233.05493 269.54195
#>  [8] 305.97068 327.38669 349.96807  13.00244  53.05493  89.54195   0.00000
#> [15]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [22]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [29]   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000   0.00000
#> [36]   0.00000   0.00000
#> 
#> $ascmc
#>  [1] 125.97068  13.00244  12.00000 263.98576 101.07060  71.50788 116.41241
#>  [8] 251.50788   0.00000   0.00000
#> 
swe_house_name('G')
#> [1] "Gauquelin sectors"